[REF] hr*: replace avatar_user fields by employee
This commit replaces fields `many2*_avatar_user` by
`many2*_avatar_employee` when it's used on `hr.employee` fields.
This allows to remove a patch on `Many2*AvatarUserField` to support
`hr.employee` features. This patch could be useless since we already
have a field with those features.
