+++ addons/hr/static/src/views/fields/many2many_avatar_employee_field/many2many_avatar_employee_field.js
@@ -2,6 +2,7 @@ import { registry } from "@web/core/registry";
 import {
     Many2ManyTagsAvatarUserField,
     KanbanMany2ManyTagsAvatarUserField,
+    ListMany2ManyTagsAvatarUserField,
     many2ManyTagsAvatarUserField,
     kanbanMany2ManyTagsAvatarUserField,
     listMany2ManyTagsAvatarUserField,
@@ -10,7 +11,14 @@ import { EmployeeFieldRelationMixin } from "@hr/views/fields/employee_field_rela
 
 export class Many2ManyTagsAvatarEmployeeField extends EmployeeFieldRelationMixin(
     Many2ManyTagsAvatarUserField
-) {}
+) {
+    displayAvatarCard(record) {
+        return (
+            (!this.env.isSmall && ["hr.employee", "hr.employee.public"].includes(this.relation)) ||
+            super.displayAvatarCard(record)
+        );
+    }
+}
 
 export const many2ManyTagsAvatarEmployeeField = {
     ...many2ManyTagsAvatarUserField,
@@ -30,7 +38,14 @@ registry.category("fields").add("many2many_avatar_employee", many2ManyTagsAvatar
 
 export class KanbanMany2ManyTagsAvatarEmployeeField extends EmployeeFieldRelationMixin(
     KanbanMany2ManyTagsAvatarUserField
-) {}
+) {
+    displayAvatarCard(record) {
+        return (
+            (!this.env.isSmall && ["hr.employee", "hr.employee.public"].includes(this.relation)) ||
+            super.displayAvatarCard(record)
+        );
+    }
+}
 
 export const kanbanMany2ManyTagsAvatarEmployeeField = {
     ...kanbanMany2ManyTagsAvatarUserField,
@@ -47,9 +62,23 @@ export const kanbanMany2ManyTagsAvatarEmployeeField = {
 
 registry
     .category("fields")
-    .add("kanban.many2many_avatar_employee", kanbanMany2ManyTagsAvatarEmployeeField);
+    .add("kanban.many2many_avatar_employee", kanbanMany2ManyTagsAvatarEmployeeField)
+    .add("activity.many2many_avatar_employee", kanbanMany2ManyTagsAvatarEmployeeField);
+
+export class ListMany2ManyTagsAvatarEmployeeField extends EmployeeFieldRelationMixin(
+    ListMany2ManyTagsAvatarUserField
+) {
+    displayAvatarCard(record) {
+        return (
+            (!this.env.isSmall && ["hr.employee", "hr.employee.public"].includes(this.relation)) ||
+            super.displayAvatarCard(record)
+        );
+    }
+}
+
 export const listMany2ManyTagsAvatarEmployeeField = {
     ...listMany2ManyTagsAvatarUserField,
+    component: ListMany2ManyTagsAvatarEmployeeField,
     additionalClasses: [
         ...listMany2ManyTagsAvatarUserField.additionalClasses,
         "o_field_many2many_avatar_user",
