+++ addons/hr_contract/report/hr_contract_history_report_views.xml
@@ -58,7 +58,7 @@
                         </div>
                     </h1>
                     <h2>
-                        <field name="employee_id" widget="many2one_avatar_user"/>
+                        <field name="employee_id" widget="many2one_avatar_employee"/>
                     </h2>
                     <group>
                         <group>
