+++ addons/hr_expense/wizard/hr_expense_approve_duplicate_views.xml
@@ -9,7 +9,7 @@
                 <field name="expense_ids" nolabel="1">
                     <list>
                         <field name="date" readonly="1" />
-                        <field name="employee_id" readonly="1" widget="many2one_avatar_user"/>
+                        <field name="employee_id" readonly="1" widget="many2one_avatar_employee"/>
                         <field name="product_id" readonly="1" />
                         <field name="total_amount" readonly="1" />
                         <field name="name" readonly="1" />
