+++ addons/hr_expense/wizard/hr_expense_split_wizard_views.xml
@@ -6,7 +6,7 @@
             <field name="model">hr.expense.split.wizard</field>
             <field name="arch" type="xml">
                 <form>
-                    <div class="alert alert-warning w-100 d-flex align-items-center gap-1" 
+                    <div class="alert alert-warning w-100 d-flex align-items-center gap-1"
                         invisible="split_possible"  role="alert">
                         <span>The total amount doesn't match the original amount.</span>
                     </div>
@@ -21,7 +21,7 @@
                             <field name="product_has_cost" column_invisible="True"/>
                             <field name="name"/>
                             <field name="product_id"/>
-                            <field name="employee_id" widget="many2one_avatar_user"/>
+                            <field name="employee_id" widget="many2one_avatar_employee"/>
                             <field name="tax_ids" widget="many2many_tax_tags" readonly="not product_has_tax"/>
                             <field name="tax_amount_currency"/>
                             <field name="analytic_distribution" widget="analytic_distribution"
