+++ addons/hr_fleet/views/fleet_vehicle_cost_views.xml
@@ -27,7 +27,7 @@
         <field name="inherit_id" ref="fleet.fleet_vehicle_log_contract_view_tree"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='purchaser_id']" position="after">
-                <field name="purchaser_employee_id" optional="hide" widget="many2one_avatar_user"/>
+                <field name="purchaser_employee_id" optional="hide" widget="many2one_avatar_employee"/>
             </xpath>
         </field>
     </record>
@@ -61,7 +61,7 @@
         <field name="inherit_id" ref="fleet.fleet_vehicle_log_services_view_tree"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='purchaser_id']" position="after">
-                <field name="purchaser_employee_id" readonly="1" widget="many2one_avatar_user" optional="hide"/>
+                <field name="purchaser_employee_id" readonly="1" widget="many2one_avatar_employee" optional="hide"/>
             </xpath>
         </field>
     </record>
