+++ addons/hr_holidays/wizard/hr_leave_allocation_generate_multi_wizard_views.xml
@@ -12,7 +12,7 @@
                 <group>
                     <group>
                         <field name="allocation_mode" string="Mode"/>
-                        <field name="employee_ids" invisible="allocation_mode != 'employee'" required="allocation_mode == 'employee'" widget="many2many_avatar_user"/>
+                        <field name="employee_ids" invisible="allocation_mode != 'employee'" required="allocation_mode == 'employee'" widget="many2many_avatar_employee"/>
                         <field name="company_id" invisible="allocation_mode != 'company'"/>
                         <field name="department_id" invisible="allocation_mode != 'department'" required="allocation_mode == 'department'"/>
                         <field name="category_id" invisible="allocation_mode != 'category'" required="allocation_mode == 'category'"/>
