+++ addons/hr_skills/report/hr_employee_skill_report_views.xml
@@ -28,7 +28,7 @@
         <field name="model">hr.employee.skill.report</field>
         <field name="arch" type="xml">
             <list expand="0" edit="1" editable="bottom">
-                <field name="employee_id" widget="many2one_avatar_user" options="{'no_create': True, 'no_open': True}"/>
+                <field name="employee_id" widget="many2one_avatar_employee" options="{'no_create': True, 'no_open': True}"/>
                 <field name="skill_type_id" options="{'no_create': True, 'no_open': True}"/>
                 <field name="skill_id" options="{'no_create': True, 'no_open': True}"/>
                 <field name="skill_level" options="{'no_create': True, 'no_open': True}"/>
