+++ addons/hr_skills_survey/views/hr_employee_certification_views.xml
@@ -7,7 +7,7 @@
                 decoration-danger="expiration_status == 'expired'"
                 decoration-warning="expiration_status == 'expiring'"
                 default_group_by="employee_id">
-                <field name="employee_id" widget="many2one_avatar_user"/>
+                <field name="employee_id" widget="many2one_avatar_employee"/>
                 <field name="name"/>
                 <field name="date_start" string="Validity Start"/>
                 <field name="date_end" string="Validity End"/>
