+++ addons/hr_work_entry/views/hr_work_entry_views.xml
@@ -40,7 +40,7 @@
                 color="color"
                 event_limit="5">
                 <!-- Sidebar favorites filters -->
-                <field name="employee_id" write_model="hr.user.work.entry.employee" write_field="employee_id" avatar_field="avatar_128" widget="many2one_avatar_user"/>
+                <field name="employee_id" write_model="hr.user.work.entry.employee" write_field="employee_id" avatar_field="avatar_128" widget="many2one_avatar_employee"/>
                 <field name="state"/>
             </calendar>
         </field>
@@ -61,7 +61,7 @@
                     <group>
                         <group>
                             <field name="name" string="Description" placeholder="Work Entry Name" readonly="state == 'validated'"/>
-                            <field name="employee_id" readonly="state != 'draft'" widget="many2one_avatar_user"/>
+                            <field name="employee_id" readonly="state != 'draft'" widget="many2one_avatar_employee"/>
                             <field name="work_entry_type_id" readonly="state == 'validated'" options="{'no_create': True, 'no_open': True}"/>
                         </group>
                         <group>
