+++ addons/hr_work_entry_contract/wizard/hr_work_entry_regeneration_wizard_views.xml
@@ -7,7 +7,7 @@
             <form string="Regenerate Employee Work Entries">
                 <group>
                     <group>
-                        <field name="employee_ids" widget="many2many_avatar_user"/>
+                        <field name="employee_ids" widget="many2many_avatar_employee"/>
                         <label for="date_from"></label>
                         <div name="date_from">
                             <div class="text-info" invisible="earliest_available_date_message == ''">
